requirements.txt
streamlit
appwrite
pandas
python-dotenv